Move DNS management methods to the LoadBalancedInstance mixin.
